cambios

+++ php/delete-user.php
@@ -1,13 +1,28 @@
 <?php
 session_start();
 require_once("../connection.php");
-$userId=$_SESSION['user']['user-id'];
-$sql=$mysqli->prepare('DELETE FROM usuarios WHERE id = ?');
-$sql->bind_param("i", $userId);
-$sql->execute();
-if(!$sql){
-    die( "Id: " . $userId . " Error: " . $sql . "<br>" . $mysqli->error);
+
+if (!isset($_SESSION['user']['user-id'])) {
+    die("User ID not set in session");
 }
+
+$userId = $_SESSION['user']['user-id'];
+
+$sql = $mysqli->prepare('DELETE FROM usuarios WHERE id = ?');
+
+if (!$sql) {
+    die("Error preparing SQL: " . $mysqli->error);
+}
+
+$sql->bind_param("i", $userId);
+
+if (!$sql->execute()) {
+    die("Error executing SQL: " . $sql->error);
+}
+
 $sql->close();
 $mysqli->close();
+
 header("Location: ../session-destroy.php");
+exit;
+?>
